docs(README.md): update README.md to reflect ongoing development of a new version of the ESLint configuration
The current version of the ESLint configuration is being updated to support ESLint's new flat config, dynamic adaptation to the project (JS, TS, React, etc), and other improvements. This update is based on the work of [Antfu](https://github.com/antfu/eslint-config).

The README.md file is updated to inform users about the ongoing development and to provide a note about the new version.